+++ apps/laboratory/src/components/AppKitButtons.tsx
@@ -58,7 +58,7 @@ export function AppKitButtons() {
             <Heading size="xs" textTransform="uppercase" pb="2">
               Hooks Interactions
             </Heading>
-            <Button data-testId="w3m-open-hook-button" onClick={openModalWithHook}>
+            <Button data-testid="w3m-open-hook-button" onClick={openModalWithHook}>
               Open
             </Button>
           </Box>