+++ apps/laboratory/tests/shared/fixtures/w3m-fixture.ts
@@ -27,7 +27,7 @@ export const testM = timingFixture.extend<ModalFixture>({
 export const testMSiwe = timingFixture.extend<ModalFixture>({
   library: ['wagmi', { option: true }],
   modalPage: async ({ page, library }, use) => {
-    const modalPage = new ModalPage(page, library, 'siwe')
+    const modalPage = new ModalPage(page, library, 'all')
     await modalPage.load()
     await use(modalPage)
   }