+++ packages/base/adapters/evm/ethers/client.ts
@@ -53,6 +53,7 @@ import type { W3mFrameTypes } from '@web3modal/wallet'
 import type { AppKit } from '../../../src/client.js'
 import type { AppKitOptions } from '../../../utils/TypesUtil.js'
 import type { OptionsControllerState } from '@web3modal/core'
+import { SafeLocalStorage } from '../../../utils/SafeLocalStorage.js'
 
 // -- Types ---------------------------------------------------------------------
 export interface AdapterOptions extends Pick<AppKitOptions, 'siweConfig'> {
@@ -168,7 +169,7 @@ export class EVMEthersClient {
 
       getApprovedCaipNetworksData: async () =>
         new Promise(async resolve => {
-          const walletChoice = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+          const walletChoice = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
           if (walletChoice?.includes(ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID)) {
             const provider = await this.getWalletConnectProvider()
             if (!provider) {
@@ -343,7 +344,7 @@ export class EVMEthersClient {
       disconnect: async () => {
         const provider = EthersStoreUtil.state.provider
         const providerType = EthersStoreUtil.state.providerType
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
         this.appKit?.setClientId(null)
         if (siweConfig?.options?.signOutOnDisconnect) {
@@ -368,7 +369,7 @@ export class EVMEthersClient {
         }
 
         provider?.emit?.('disconnect')
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
       },
 
@@ -565,7 +566,7 @@ export class EVMEthersClient {
       this.checkActiveInjectedProvider(this.ethersConfig)
     }
 
-    if (this.ethersConfig.auth) {
+    if (this.ethersConfig.auth?.email || this.ethersConfig.auth?.socials?.length) {
       this.syncAuthConnector(this.options.projectId, this.ethersConfig.auth)
     }
 
@@ -641,7 +642,7 @@ export class EVMEthersClient {
   public async disconnect() {
     const { provider, providerType } = EthersStoreUtil.state
 
-    localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+    SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
     EthersStoreUtil.reset()
     this.appKit?.setClientId(null)
 
@@ -752,7 +753,7 @@ export class EVMEthersClient {
 
   private async checkActiveWalletConnectProvider() {
     const WalletConnectProvider = await this.getWalletConnectProvider()
-    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const walletId = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
 
     if (WalletConnectProvider) {
       if (walletId === ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID) {
@@ -766,7 +767,7 @@ export class EVMEthersClient {
 
   private checkActiveInjectedProvider(config: ProviderType) {
     const InjectedProvider = config.injected
-    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const walletId = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
 
     if (InjectedProvider) {
       if (walletId === ConstantsUtil.INJECTED_CONNECTOR_ID) {
@@ -778,7 +779,7 @@ export class EVMEthersClient {
 
   private checkActiveCoinbaseProvider(config: ProviderType) {
     const CoinbaseProvider = config.coinbase as unknown as ExternalProvider
-    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const walletId = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
 
     if (CoinbaseProvider) {
       if (walletId === ConstantsUtil.COINBASE_SDK_CONNECTOR_ID) {
@@ -786,7 +787,7 @@ export class EVMEthersClient {
           this.setCoinbaseProvider(config)
           this.watchCoinbase(config)
         } else {
-          localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+          SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
           EthersStoreUtil.reset()
         }
       }
@@ -794,7 +795,7 @@ export class EVMEthersClient {
   }
 
   private checkActive6963Provider() {
-    const currentActiveWallet = window?.localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const currentActiveWallet = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
     if (currentActiveWallet) {
       const currentProvider = this.EIP6963Providers.find(
         provider => provider.info.name === currentActiveWallet
@@ -806,7 +807,7 @@ export class EVMEthersClient {
   }
 
   private async setWalletConnectProvider() {
-    window?.localStorage.setItem(
+    SafeLocalStorage.setItem(
       EthersConstantsUtil.WALLET_ID,
       ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID
     )
@@ -817,6 +818,7 @@ export class EVMEthersClient {
       EthersStoreUtil.setProvider(WalletConnectProvider as unknown as Provider)
       EthersStoreUtil.setStatus('connected')
       EthersStoreUtil.setIsConnected(true)
+
       this.appKit?.setAllAccounts(
         WalletConnectProvider.accounts.map(address => ({ address, type: 'eoa' })),
         this.chain
@@ -834,7 +836,7 @@ export class EVMEthersClient {
   }
 
   private async setInjectedProvider(config: ProviderType) {
-    window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.INJECTED_CONNECTOR_ID)
+    SafeLocalStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.INJECTED_CONNECTOR_ID)
     const InjectedProvider = config.injected
 
     if (InjectedProvider) {
@@ -856,7 +858,7 @@ export class EVMEthersClient {
   }
 
   private async setEIP6963Provider(provider: Provider, name: string) {
-    window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, name)
+    SafeLocalStorage.setItem(EthersConstantsUtil.WALLET_ID, name)
 
     if (provider) {
       const { addresses, chainId } = await EthersHelpersUtil.getUserInfo(provider)
@@ -877,10 +879,7 @@ export class EVMEthersClient {
   }
 
   private async setCoinbaseProvider(config: ProviderType) {
-    window?.localStorage.setItem(
-      EthersConstantsUtil.WALLET_ID,
-      ConstantsUtil.COINBASE_SDK_CONNECTOR_ID
-    )
+    SafeLocalStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.COINBASE_SDK_CONNECTOR_ID)
     const CoinbaseProvider = config.coinbase
     if (CoinbaseProvider) {
       const { addresses, chainId } = await EthersHelpersUtil.getUserInfo(CoinbaseProvider)
@@ -901,7 +900,7 @@ export class EVMEthersClient {
   }
 
   private async setAuthProvider() {
-    window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.AUTH_CONNECTOR_ID)
+    SafeLocalStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.AUTH_CONNECTOR_ID)
 
     if (this.authProvider) {
       this.appKit?.setLoading(true)
@@ -944,7 +943,7 @@ export class EVMEthersClient {
     const provider = await this.getWalletConnectProvider()
 
     function disconnectHandler() {
-      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+      SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
       EthersStoreUtil.reset()
       provider?.removeListener('disconnect', disconnectHandler)
       provider?.removeListener('accountsChanged', accountsChangedHandler)
@@ -975,7 +974,7 @@ export class EVMEthersClient {
     const provider = config.injected
 
     function disconnectHandler() {
-      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+      SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
       EthersStoreUtil.reset()
 
       provider?.removeListener('disconnect', disconnectHandler)
@@ -988,7 +987,7 @@ export class EVMEthersClient {
       if (currentAccount) {
         EthersStoreUtil.setAddress(getOriginalAddress(currentAccount) as Address)
       } else {
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
       }
     }
@@ -1012,7 +1011,7 @@ export class EVMEthersClient {
 
   private watchEIP6963(provider: Provider) {
     function disconnectHandler() {
-      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+      SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
       EthersStoreUtil.reset()
       provider.removeListener('disconnect', disconnectHandler)
       provider.removeListener('accountsChanged', accountsChangedHandler)
@@ -1029,7 +1028,7 @@ export class EVMEthersClient {
         )
       } else {
         this.appKit?.setAllAccounts([], this.chain)
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
       }
     }
@@ -1053,9 +1052,9 @@ export class EVMEthersClient {
 
   private watchCoinbase(config: ProviderType) {
     const provider = config.coinbase
-    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const walletId = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
     function disconnectHandler() {
-      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+      SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
       EthersStoreUtil.reset()
 
       provider?.removeListener('disconnect', disconnectHandler)
@@ -1068,7 +1067,7 @@ export class EVMEthersClient {
       if (currentAccount) {
         EthersStoreUtil.setAddress(getOriginalAddress(currentAccount) as Address)
       } else {
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
       }
     }
@@ -1321,7 +1320,7 @@ export class EVMEthersClient {
   }
 
   private syncConnectedWalletInfo() {
-    const currentActiveWallet = window?.localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const currentActiveWallet = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
     const providerType = EthersStoreUtil.state.providerType
 
     if (providerType === ConstantsUtil.EIP6963_CONNECTOR_ID) {
@@ -1336,7 +1335,6 @@ export class EVMEthersClient {
       }
     } else if (providerType === ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID) {
       const provider = EthersStoreUtil.state.provider as unknown as EthereumProvider
-
       if (provider.session) {
         this.appKit?.setConnectedWalletInfo(
           {
@@ -1347,6 +1345,18 @@ export class EVMEthersClient {
           this.chain
         )
       }
+    } else if (providerType === ConstantsUtil.COINBASE_SDK_CONNECTOR_ID) {
+      const connector = this.appKit
+        ?.getConnectors()
+        .find(c => c.id === ConstantsUtil.COINBASE_SDK_CONNECTOR_ID)
+
+      this.appKit?.setConnectedWalletInfo(
+        {
+          name: 'Coinbase Wallet',
+          icon: this.appKit?.getConnectorImage(connector)
+        },
+        this.chain
+      )
     } else if (currentActiveWallet) {
       this.appKit?.setConnectedWalletInfo({ name: currentActiveWallet }, this.chain)
     }
@@ -1556,15 +1566,8 @@ export class EVMEthersClient {
       const { info, provider } = event.detail
       const connectors = this.appKit?.getConnectors()
       const existingConnector = connectors?.find(c => c.name === info.name)
-      const coinbaseConnector = connectors?.find(
-        c => c.id === ConstantsUtil.COINBASE_SDK_CONNECTOR_ID
-      )
-      const isCoinbaseDuplicated =
-        coinbaseConnector &&
-        event.detail.info.rdns ===
-          ConstantsUtil.CONNECTOR_RDNS_MAP[ConstantsUtil.COINBASE_SDK_CONNECTOR_ID]
 
-      if (!existingConnector && !isCoinbaseDuplicated) {
+      if (!existingConnector) {
         const type = PresetsUtil.ConnectorTypesMap[ConstantsUtil.EIP6963_CONNECTOR_ID]
         if (type) {
           this.appKit?.addConnector({