+++ packages/base/adapters/evm/ethers5/client.ts
@@ -37,6 +37,7 @@ import type { Chain as AvailableChain } from '@web3modal/common'
 import type { AppKit } from '../../../src/client.js'
 import type { AppKitOptions } from '../../../utils/TypesUtil.js'
 import type { OptionsControllerState } from '@web3modal/core'
+import { SafeLocalStorage } from '../../../utils/SafeLocalStorage.js'
 
 // -- Types ---------------------------------------------------------------------
 export interface AdapterOptions extends Pick<AppKitOptions, 'siweConfig'> {
@@ -150,7 +151,7 @@ export class EVMEthers5Client {
 
       getApprovedCaipNetworksData: async () =>
         new Promise(async resolve => {
-          const walletChoice = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+          const walletChoice = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
           if (walletChoice?.includes(ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID)) {
             const provider = await this.getWalletConnectProvider()
             if (!provider) {
@@ -317,7 +318,7 @@ export class EVMEthers5Client {
       disconnect: async () => {
         const provider = EthersStoreUtil.state.provider
         const providerType = EthersStoreUtil.state.providerType
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
         this.appKit?.setClientId(null)
         if (siweConfig?.options?.signOutOnDisconnect) {
@@ -340,7 +341,7 @@ export class EVMEthers5Client {
         }
 
         provider?.emit?.('disconnect')
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
       },
 
@@ -522,7 +523,7 @@ export class EVMEthers5Client {
 
   public async disconnect() {
     const { provider, providerType } = EthersStoreUtil.state
-    localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+    SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
     EthersStoreUtil.reset()
 
     this.appKit?.setClientId(null)
@@ -599,7 +600,7 @@ export class EVMEthers5Client {
 
   private async checkActiveWalletConnectProvider() {
     const WalletConnectProvider = await this.getWalletConnectProvider()
-    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const walletId = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
 
     if (WalletConnectProvider) {
       if (walletId === ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID) {
@@ -613,7 +614,7 @@ export class EVMEthers5Client {
 
   private checkActiveInjectedProvider(config: ProviderType) {
     const InjectedProvider = config.injected
-    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const walletId = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
 
     if (InjectedProvider) {
       if (walletId === ConstantsUtil.INJECTED_CONNECTOR_ID) {
@@ -625,7 +626,7 @@ export class EVMEthers5Client {
 
   private checkActiveCoinbaseProvider(config: ProviderType) {
     const CoinbaseProvider = config.coinbase as unknown as ExternalProvider
-    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const walletId = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
 
     if (CoinbaseProvider) {
       if (walletId === ConstantsUtil.COINBASE_SDK_CONNECTOR_ID) {
@@ -633,7 +634,7 @@ export class EVMEthers5Client {
           this.setCoinbaseProvider(config)
           this.watchCoinbase(config)
         } else {
-          localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+          SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
           EthersStoreUtil.reset()
         }
       }
@@ -641,7 +642,7 @@ export class EVMEthers5Client {
   }
 
   private checkActive6963Provider() {
-    const currentActiveWallet = window?.localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const currentActiveWallet = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
     if (currentActiveWallet) {
       const currentProvider = this.EIP6963Providers.find(
         provider => provider.info.name === currentActiveWallet
@@ -653,7 +654,7 @@ export class EVMEthers5Client {
   }
 
   private async setWalletConnectProvider() {
-    window?.localStorage.setItem(
+    SafeLocalStorage.setItem(
       EthersConstantsUtil.WALLET_ID,
       ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID
     )
@@ -674,7 +675,7 @@ export class EVMEthers5Client {
   }
 
   private async setEIP6963Provider(provider: Provider, name: string) {
-    window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, name)
+    SafeLocalStorage.setItem(EthersConstantsUtil.WALLET_ID, name)
 
     if (provider) {
       const { addresses, chainId } = await EthersHelpersUtil.getUserInfo(provider)
@@ -695,7 +696,7 @@ export class EVMEthers5Client {
   }
 
   private async setInjectedProvider(config: ProviderType) {
-    window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.INJECTED_CONNECTOR_ID)
+    SafeLocalStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.INJECTED_CONNECTOR_ID)
     const InjectedProvider = config.injected
 
     if (InjectedProvider) {
@@ -717,10 +718,7 @@ export class EVMEthers5Client {
   }
 
   private async setCoinbaseProvider(config: ProviderType) {
-    window?.localStorage.setItem(
-      EthersConstantsUtil.WALLET_ID,
-      ConstantsUtil.COINBASE_SDK_CONNECTOR_ID
-    )
+    SafeLocalStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.COINBASE_SDK_CONNECTOR_ID)
     const CoinbaseProvider = config.coinbase
     if (CoinbaseProvider) {
       const { addresses, chainId } = await EthersHelpersUtil.getUserInfo(CoinbaseProvider)
@@ -784,7 +782,7 @@ export class EVMEthers5Client {
     const WalletConnectProvider = await this.getWalletConnectProvider()
 
     function disconnectHandler() {
-      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+      SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
       EthersStoreUtil.reset()
 
       WalletConnectProvider?.removeListener('disconnect', disconnectHandler)
@@ -816,7 +814,7 @@ export class EVMEthers5Client {
     const InjectedProvider = config.injected
 
     function disconnectHandler() {
-      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+      SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
       EthersStoreUtil.reset()
 
       InjectedProvider?.removeListener('disconnect', disconnectHandler)
@@ -829,7 +827,7 @@ export class EVMEthers5Client {
       if (currentAccount) {
         EthersStoreUtil.setAddress(utils.getAddress(currentAccount) as Address)
       } else {
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
       }
     }
@@ -856,7 +854,7 @@ export class EVMEthers5Client {
     const namespace = this.chain
 
     function disconnectHandler() {
-      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+      SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
       EthersStoreUtil.reset()
 
       provider.removeListener('disconnect', disconnectHandler)
@@ -874,7 +872,7 @@ export class EVMEthers5Client {
         )
       } else {
         appKit?.setAllAccounts([], namespace)
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
       }
     }
@@ -896,10 +894,10 @@ export class EVMEthers5Client {
 
   private watchCoinbase(config: ProviderType) {
     const CoinbaseProvider = config.coinbase
-    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const walletId = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
 
     function disconnectHandler() {
-      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+      SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
       EthersStoreUtil.reset()
 
       CoinbaseProvider?.removeListener('disconnect', disconnectHandler)
@@ -909,7 +907,7 @@ export class EVMEthers5Client {
 
     function accountsChangedHandler(accounts: string[]) {
       if (accounts.length === 0) {
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        SafeLocalStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
       } else {
         EthersStoreUtil.setAddress(accounts[0] as Address)
@@ -1155,7 +1153,7 @@ export class EVMEthers5Client {
   }
 
   private syncConnectedWalletInfo() {
-    const currentActiveWallet = window?.localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+    const currentActiveWallet = SafeLocalStorage.getItem(EthersConstantsUtil.WALLET_ID)
     const providerType = EthersStoreUtil.state.providerType
 
     if (providerType === ConstantsUtil.EIP6963_CONNECTOR_ID) {
@@ -1390,15 +1388,8 @@ export class EVMEthers5Client {
       const { info, provider } = event.detail
       const connectors = this.appKit?.getConnectors()
       const existingConnector = connectors?.find(c => c.name === info.name)
-      const coinbaseConnector = connectors?.find(
-        c => c.id === ConstantsUtil.COINBASE_SDK_CONNECTOR_ID
-      )
-      const isCoinbaseDuplicated =
-        coinbaseConnector &&
-        event.detail.info.rdns ===
-          ConstantsUtil.CONNECTOR_RDNS_MAP[ConstantsUtil.COINBASE_SDK_CONNECTOR_ID]
 
-      if (!existingConnector && !isCoinbaseDuplicated) {
+      if (!existingConnector) {
         const type = PresetsUtil.ConnectorTypesMap[ConstantsUtil.EIP6963_CONNECTOR_ID]
         if (type) {
           this.appKit?.addConnector({